ENH: Separate general bits of tracer analysis and create a new base class. Add more safety checks to tracer analysis.
